Refactor admin search and health endpoints into dedicated router
