+++ routers/api_admin.py
@@ -1,0 +1,107 @@
+"""Admin/search/health endpoints split from the main API router."""
+
+from fastapi import APIRouter, Depends, Query
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlmodel import select, func
+from typing import List
+
+from core.database import get_session
+from core.security import get_current_username
+from models import Product, ProductTagLink, Service, Tag, TagGroup
+
+router = APIRouter(tags=["api"])
+
+
+@router.get("/products/search")
+async def search_products(
+    q: str = None,
+    is_inverter: bool = None,
+    session: AsyncSession = Depends(get_session),
+):
+    """Search products with fuzzy matching."""
+    from services.product_service import ProductService
+
+    products = await ProductService.search(session, query=q, is_inverter=is_inverter)
+    return {"items": products}
+
+
+@router.get("/admin/tags/filterable")
+async def get_filterable_tags(
+    session: AsyncSession = Depends(get_session),
+    username: str = Depends(get_current_username),
+):
+    stmt = (
+        select(TagGroup, Tag)
+        .join(Tag, Tag.group_id == TagGroup.id)
+        .where(Tag.is_filter == True)
+        .order_by(TagGroup.sort_order, TagGroup.title, Tag.sort_order, Tag.title)
+    )
+    result = await session.execute(stmt)
+
+    grouped = {}
+    for row in result:
+        group, tag = row
+        if group.title not in grouped:
+            grouped[group.title] = {
+                "group_label": group.title,
+                "tags": [],
+            }
+        grouped[group.title]["tags"].append(
+            {
+                "id": tag.id,
+                "title": tag.title,
+                "slug": tag.slug,
+            }
+        )
+
+    return list(grouped.values())
+
+
+@router.get("/admin/products/search")
+async def admin_search_products(
+    q: str = "",
+    tag_ids: List[int] = Query(None),
+    session: AsyncSession = Depends(get_session),
+    username: str = Depends(get_current_username),
+):
+    stmt = select(Product)
+
+    if tag_ids:
+        tag_subquery = (
+            select(ProductTagLink.product_id)
+            .where(ProductTagLink.tag_id.in_(tag_ids))
+            .group_by(ProductTagLink.product_id)
+            .having(func.count(ProductTagLink.tag_id) == len(tag_ids))
+        )
+        stmt = stmt.where(Product.id.in_(tag_subquery))
+
+    if q:
+        pattern = f"%{q}%"
+        stmt = stmt.where((Product.title.ilike(pattern)) | (Product.description.ilike(pattern)))
+
+    stmt = stmt.limit(20)
+    result = await session.execute(stmt)
+    products = result.scalars().all()
+    return [{"id": p.id, "text": p.title, "price": p.price} for p in products]
+
+
+@router.get("/admin/services/search")
+async def admin_search_services(
+    q: str = "",
+    session: AsyncSession = Depends(get_session),
+    username: str = Depends(get_current_username),
+):
+    stmt = select(Service).where(Service.title.ilike(f"%{q}%")).limit(20)
+    result = await session.execute(stmt)
+    services = result.scalars().all()
+    return [{"id": s.id, "text": s.title, "price": s.base_price} for s in services]
+
+
+@router.get("/health")
+async def health_check(session: AsyncSession = Depends(get_session)):
+    """Check API and database availability."""
+    try:
+        await session.execute(select(1))
+        return {"status": "ok", "database": "online"}
+    except Exception as e:
+        return {"status": "error", "database": "offline", "detail": str(e)}
